# ipx/docs/swagger.yaml
basePath: /
definitions:
...
  v2.BuildERC20EIP1559TransactionRequest:
    properties:
      amount:
        example: "1000000000000000000"
        type: string
      contract:
        example: 0x5FbDB2315678afecb367f032d93F642f64180aa3
        type: string
      from:
        example: 0xDa70aA79f1a329719b9cf9d334b0a82b1d5269f3
        type: string
      to:
        example: 0xDa70aA79f1a329719b9cf9d334b0a82b1d5269f3
        type: string
    type: object
  v2.BuildERC20EIP1559TransactionResponse:
    properties:
      unsigned_rlp:
        type: string
    type: object
  v2.BuildERC20LegacyTransactionRequest:
    properties:
      amount:
        example: "1000000000000000000"
        type: string
      contract:
        example: 0x5FbDB2315678afecb367f032d93F642f64180aa3
        type: string
      from:
        example: 0xDa70aA79f1a329719b9cf9d334b0a82b1d5269f3
        type: string
      to:
        example: 0xDa70aA79f1a329719b9cf9d334b0a82b1d5269f3
        type: string
    type: object
  v2.BuildERC20LegacyTransactionResponse:
    properties:
      unsigned_rlp:
        type: string
    type: object
  v2.BuildNativeEIP1559TransactionRequest:
    properties:
      from:
        type: string
      to:
        type: string
      value:
        type: string
    type: object
  v2.BuildNativeEIP1559TransactionResponse:
    properties:
      unsigned_rlp:
        type: string
    type: object
  v2.BuildNativeLegacyTransactionRequest:
    properties:
      from:
        type: string
      to:
        type: string
      value:
        type: string
    type: object
  v2.BuildNativeLegacyTransactionResponse:
    properties:
      unsigned_rlp:
        type: string
    type: object
...
host: localhost:33152
info:
  contact: {}
  description: Ethereum transaction API
  title: evmlab API
  version: "1.0"
paths:
...
  /evm/v2/transaction/erc20/eip1559:
    post:
      consumes:
      - application/json
      description: Builds transfer(address,uint256) calldata internally and returns
        the unsigned RLP-encoded EIP-1559 transaction. Gas limit is estimated on-chain
        with a 20% buffer.
      parameters:
      - description: ERC-20 transfer request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/v2.BuildERC20EIP1559TransactionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v2.BuildERC20EIP1559TransactionResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "502":
          description: Bad Gateway
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Build unsigned ERC-20 EIP-1559 transfer tx
      tags:
      - transfer
  /evm/v2/transaction/erc20/legacy:
    post:
      consumes:
      - application/json
      description: Builds transfer(address,uint256) calldata internally and returns
        the unsigned RLP-encoded EIP-155 legacy transaction. Gas limit is estimated
        on-chain with a 20% buffer.
      parameters:
      - description: ERC-20 transfer request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/v2.BuildERC20LegacyTransactionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v2.BuildERC20LegacyTransactionResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "502":
          description: Bad Gateway
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Build unsigned ERC-20 legacy transfer tx
      tags:
      - transfer
  /evm/v2/transaction/native/eip1559:
    post:
      consumes:
      - application/json
      description: Fetches chain state (chainID, nonce, fees) and returns the unsigned
        RLP-encoded EIP-1559 transfer transaction
      parameters:
      - description: BuildNativeEIP1559Transaction request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/v2.BuildNativeEIP1559TransactionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v2.BuildNativeEIP1559TransactionResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "502":
          description: Bad Gateway
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Build unsigned EIP-1559 transfer tx
      tags:
      - transfer
  /evm/v2/transaction/native/legacy:
    post:
      consumes:
      - application/json
      description: Fetches chain state (chainID, nonce, gas price) and returns the
        unsigned RLP-encoded EIP-155 legacy transfer transaction
      parameters:
      - description: Transfer request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/v2.BuildNativeLegacyTransactionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v2.BuildNativeLegacyTransactionResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "502":
          description: Bad Gateway
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Build unsigned legacy transfer tx
      tags:
      - transfer
...
swagger: "2.0"